packages/genro-traefik/pyproject.toml
# Copyright 2025 Softwell S.r.l. - Licensed under Apache License 2.0
# See LICENSE file for details

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "genro-traefik"
version = "0.1.0"
description = "Traefik v3 configuration builder for Genropy - validated YAML generation"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "Genropy Team", email = "info@genropy.org" }
]
keywords = [
    "genropy",
    "traefik",
    "reverse-proxy",
    "load-balancer",
    "configuration",
    "yaml",
    "infrastructure",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: System :: Networking",
    "Typing :: Typed",
]
dependencies = [
    "genro-bag>=0.1.0",
    "genro-scriba>=0.1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-sugar>=0.9",
    "mypy>=1.0",
    "ruff>=0.1.0",
]
docs = [
    "sphinx>=8.0.0",
    "sphinx-rtd-theme>=3.0.0",
    "sphinx-autodoc-typehints>=3.0.0",
    "myst-parser>=4.0.0",
    "sphinxcontrib-mermaid>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/genropy/genro-traefik"
Documentation = "https://genro-traefik.readthedocs.io"
Repository = "https://github.com/genropy/genro-traefik"
Issues = "https://github.com/genropy/genro-traefik/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/genro_traefik"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
strict = true
packages = ["genro_traefik"]
mypy_path = "src"

[[tool.mypy.overrides]]
module = [
    "genro_bag",
    "genro_bag.*",
    "yaml",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
# Builder has ~150 @element methods with body `...` — mypy strict
# is not applicable to declarative builder patterns
module = "genro_traefik.builders.*"
ignore_errors = true

[[tool.mypy.overrides]]
# Compiler and App use Any heavily for BagNode traversal
module = ["genro_traefik.traefik_compiler", "genro_traefik.traefik_app"]
warn_return_any = false

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "T10",    # flake8-debugger
    "ISC",    # flake8-implicit-str-concat
    "PIE",    # flake8-pie
    "PT",     # flake8-pytest-style
    "RSE",    # flake8-raise
    "RET",    # flake8-return
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "ARG",    # flake8-unused-arguments
    "PL",     # Pylint
    "RUF",    # Ruff-specific rules
]
ignore = [
    "PLR0913",  # Too many arguments
    "PLR2004",  # Magic value comparison
    "PIE790",   # Unnecessary `...` literal (needed for @element decorators)
    "ARG001",   # Unused function argument (kwargs in @element methods)
    "ARG002",   # Unused method argument (kwargs in @element methods)
    "PLC0415",  # import not at top level (yaml lazy import)
    "PLR0911",  # Too many return statements
    "N802",     # camelCase method names (Traefik convention)
    "N803",     # camelCase argument names (Traefik convention)
]

[tool.ruff.lint.per-file-ignores]
# Builder has long @element decorator lines by design
"src/genro_traefik/builders/traefik_builder.py" = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["genro_traefik"]

[tool.coverage.run]
source = ["src/genro_traefik"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
